Refactor testbench
