+++ packages/devtools/devtools-testbench/src/devtools/index.html
@@ -1,0 +1,26 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width,initial-scale=1">
+    <title>Kai</title>
+    <meta name="description" content="DXOS Composer Application">
+
+    <link rel="icon" href="/favicon.ico">
+    <meta name="theme-color" content="#ffffff">
+    <script>
+      // TODO(wittjosiah): Uncomment when ready for theming.
+      // function setTheme(darkMode) {
+      //   document.documentElement.classList[darkMode ? 'add' : 'remove']('dark')
+      // }
+      // setTheme(window.matchMedia('(prefers-color-scheme: dark)').matches)
+      // window.matchMedia('(prefers-color-scheme: dark)').addEventListener('change', function (e) {
+      //   setTheme(e.matches)
+      // });
+    </script>
+  </head>
+  <body>
+    <div id="root"></div>
+    <script type="module" src="/src/devtools/main.tsx"></script>
+  </body>
+</html>
